Dry up serialization logic for stores that need serialization
